child_exec.cを削除
